requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7